feat: 게시판 접근권한설정을 위한 permission.py 작성 및 불필요한 파일 삭제 - 유혜선
